[#139] Add rx/tx pin choosing, simple version of Morse main
